+++ system/thread-interactor.go
@@ -18,6 +18,10 @@ type ThreadRepository interface {
 	DeleteById(id uuid.UUID) error
 }
 
+type StringDeleter interface {
+	DeleteAllByThread(threadId uuid.UUID) error
+}
+
 func (t *ThreadInteractor) FindAll() ([]core.Thread, error) {
 	return t.Repo.FindAll()
 }
@@ -26,6 +30,12 @@ func (t *ThreadInteractor) CreateOne(thread core.Thread) (core.Thread, error) {
 	return t.Repo.CreateOne(thread)
 }
 
+// DeleteById deletes all the strings associated with a thread and then deletes
+// the thread itself.
 func (t *ThreadInteractor) DeleteById(id uuid.UUID) error {
+	err := t.StringDeleter.DeleteAllByThread(id)
+	if err != nil {
+		return err
+	}
 	return t.Repo.DeleteById(id)
 }