+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [2022-07-11]
+### Updated
+- delete thread now deletes all associated strings first
+- fetch strings by thread in order
+
+### Fixed
+- string FindAll wasn't using the right function to get the query parameter for `thread` id
+
 ## [2022-07-09]
 ### Added
 - createOne/deleteById/findAll strings